requirements.txt
streamlit
streamlit-image-comparison
pillow
torch
transformers
numpy
opencv-python


